Extract GitCmd for easier mocking.
